sorting fields to keep the same order always and fixed group component tests'
