+++ src/frontend/tests/onlyFront/group.spec.ts
@@ -1,50 +1,9 @@
 import { expect, test } from "@playwright/test";
 import { readFileSync } from "fs";
 
-test.describe("drag and drop test", () => {
+test.describe("group node test", () => {
   /// <reference lib="dom"/>
-  test("drop collection", async ({ page }) => {
-    await page.routeFromHAR("harFiles/langflow.har", {
-      url: "**/api/v1/**",
-      update: false,
-    });
-    await page.route("**/api/v1/flows/", async (route) => {
-      const json = {
-        id: "e9ac1bdc-429b-475d-ac03-d26f9a2a3210",
-      };
-      await route.fulfill({ json, status: 201 });
-    });
-    await page.goto("http:localhost:3000/");
-    await page.locator("span").filter({ hasText: "My Collection" }).isVisible();
-    // Read your file into a buffer.
-    const jsonContent = readFileSync(
-      "tests/onlyFront/assets/collection.json",
-      "utf-8"
-    );
-
-    // Create the DataTransfer and File
-    const dataTransfer = await page.evaluateHandle((data) => {
-      const dt = new DataTransfer();
-      // Convert the buffer to a hex array
-      const file = new File([data], "collection.json", {
-        type: "application/json",
-      });
-      dt.items.add(file);
-      return dt;
-    }, jsonContent);
-
-    // Now dispatch
-    await page.dispatchEvent('//*[@id="root"]/div/div[2]/div[2]', "drop", {
-      dataTransfer,
-    });
-    expect(
-      await page
-        .locator(".main-page-flows-display")
-        .evaluate((el) => el.children)
-    ).toBeTruthy();
-  });
-
-  test("drop flow", async ({ page }) => {
+  test("group and ungroup updating values", async ({ page }) => {
     await page.routeFromHAR("harFiles/langflow.har", {
       url: "**/api/v1/**",
       update: false,
@@ -160,10 +119,11 @@ test.describe("drag and drop test", () => {
       )
       .click();
     await page.getByLabel("Ungroup").click();
-    await expect(page.locator("#input-0")).toHaveValue("fieldValue");
+    await expect(page.locator('//*[@id="input-2"]')).toHaveValue("fieldValue");
     expect(
       await page
-        .locator('//*[@id="headlessui-listbox-button-:r4a:"]/span[1]')
+        .getByTestId(/.*rf__node-AgentInitializer.*/)
+        .getByRole("button", { name: "openai-functions" })
         .textContent()
     ).toBe("openai-functions");
   });
